fix:This is the client note:
It seems the problem arises when you click to Sponsored calendar, then the events do not come up. If you switch to Coffee related or art related calendar after that it doesn't either. However, if you go to main home page and click Calendartab at the top, the events will be shown on the main calendar format view and the summarised event for month view as per our inputs.

slug was appended with a code. so i amended the query to search with name and type instead of slug.